requirements.txt
telethon==1.29.0
selenium==4.11.2
webdriver-manager==4.0.0
python-dotenv==1.0.0
requests==2.31.0
